+++ src/App/App.js
@@ -1,11 +1,7 @@
 import React from 'react';
-import DogPen from '../components/DogPen/DogPen';
-import dogData from '../helpers/data/dogsData';
-import employeeData from '../helpers/data/employeesData';
-
+import Home from '../components/Home/Home';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.scss';
-import StaffRoom from '../components/StaffRoom/StaffRoom';
 
 
 class App extends React.Component {
@@ -14,17 +10,11 @@ class App extends React.Component {
     employees: [],
   }
 
-  componentDidMount() {
-    const dogs = dogData.getAllDogs();
-    const employees = employeeData.getAllEmployees();
-    this.setState({ dogs, employees });
-  }
 
   render() {
     return (
       <div className="App">
-        <DogPen dogs={this.state.dogs} />
-        <StaffRoom employees={this.state.employees} />
+      <Home />
       </div>
     );
   }
